requirements.txt
requests
beautifulsoup4
anthropic
python-dotenv
playwright
Pillow
